chore(cli): lazy-load command implementations to reduce startup time
- Extract command bodies from program.ts into sibling files (openCommands, installCommands, testCommand, etc.)
- Use lazy require() in action callbacks so each command only loads its dependencies when invoked
- Make trace CLI sub-command actions lazy-load their implementations
- Split zod into a standalone bundle (zodBundle) separate from mcpBundle so validators don't pull in the full MCP SDK
