+++ packages/playwright-core/ThirdPartyNotices.txt
@@ -137,6 +137,7 @@ This project incorporates components from the projects listed below. The origina
 -	yazl@2.5.1 (https://github.com/thejoshwolfe/yazl)
 -	zod-to-json-schema@3.25.1 (https://github.com/StefanTerdell/zod-to-json-schema)
 -	zod@4.3.5 (https://github.com/colinhacks/zod)
+-	zod@4.3.6 (https://github.com/colinhacks/zod)
 
 %% @hono/node-server@1.19.11 NOTICES AND INFORMATION BEGIN HERE
 =========================================
@@ -3545,8 +3546,34 @@ SOFTWARE.
 =========================================
 END OF zod@4.3.5 AND INFORMATION
 
+%% zod@4.3.6 NOTICES AND INFORMATION BEGIN HERE
+=========================================
+MIT License
+
+Copyright (c) 2025 Colin McDonnell
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+=========================================
+END OF zod@4.3.6 AND INFORMATION
+
 SUMMARY BEGIN HERE
 =========================================
-Total Packages: 133
+Total Packages: 134
 =========================================
 END OF SUMMARY
