+++ packages/playwright-core/bundles/zod/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "zod-bundle",
+  "version": "0.0.1",
+  "private": true,
+  "dependencies": {
+    "zod": "^4.3.5"
+  }
+}
