+++ packages/playwright-core/package.json
@@ -27,6 +27,7 @@
     "./lib/tools/cli-client/program": "./lib/tools/cli-client/program.js",
     "./lib/tools/mcp/program": "./lib/tools/mcp/program.js",
     "./lib/mcpBundle": "./lib/mcpBundle.js",
+    "./lib/zodBundle": "./lib/zodBundle.js",
     "./lib/tools/exports": "./lib/tools/exports.js",
     "./lib/remote/playwrightServer": "./lib/remote/playwrightServer.js",
     "./lib/server": "./lib/server/index.js",
