+++ packages/playwright-core/src/mcpBundle.ts
@@ -31,7 +31,6 @@ const ListRootsRequestSchema: typeof import('@modelcontextprotocol/sdk/types.js'
 const ProgressNotificationSchema: typeof import('@modelcontextprotocol/sdk/types.js').ProgressNotificationSchema = bundle.ProgressNotificationSchema;
 const ListToolsRequestSchema: typeof import('@modelcontextprotocol/sdk/types.js').ListToolsRequestSchema = bundle.ListToolsRequestSchema;
 const PingRequestSchema: typeof import('@modelcontextprotocol/sdk/types.js').PingRequestSchema = bundle.PingRequestSchema;
-const z: typeof import('zod') = bundle.z;
 
 export {
   zodToJsonSchema,
@@ -48,5 +47,4 @@ export {
   ListToolsRequestSchema,
   PingRequestSchema,
   ProgressNotificationSchema,
-  z,
 };
