+++ packages/playwright-core/src/tools/backend/common.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { z } from '../../mcpBundle';
+import { z } from '../../zodBundle';
 import { defineTabTool, defineTool } from './tool';
 import { renderTabsMarkdown } from './response';
