+++ packages/playwright-core/src/tools/backend/console.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { z } from '../../mcpBundle';
+import { z } from '../../zodBundle';
 import { defineTabTool } from './tool';
 
 const console = defineTabTool({
