+++ packages/playwright-core/src/tools/backend/dialogs.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { z } from '../../mcpBundle';
+import { z } from '../../zodBundle';
 import { defineTabTool } from './tool';
 
 export const handleDialog = defineTabTool({
