+++ packages/playwright-core/src/tools/backend/keyboard.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { z } from '../../mcpBundle';
+import { z } from '../../zodBundle';
 import { defineTabTool } from './tool';
 import { elementSchema } from './snapshot';
