+++ packages/playwright-core/src/tools/backend/pdf.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { z } from '../../mcpBundle';
+import { z } from '../../zodBundle';
 import { formatObject } from '../../utils/isomorphic/stringUtils';
 
 import { defineTabTool } from './tool';
