+++ packages/playwright-core/src/tools/backend/tabs.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { z } from '../../mcpBundle';
+import { z } from '../../zodBundle';
 import { defineTool } from './tool';
 import { renderTabsMarkdown } from './response';
