+++ packages/playwright-core/src/tools/cli-daemon/command.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { z } from '../../mcpBundle';
+import { z } from '../../zodBundle';
 
 import type zodType from 'zod';
