+++ packages/playwright-core/src/tools/cli-daemon/helpGenerator.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { z } from '../../mcpBundle';
+import { z } from '../../zodBundle';
 
 import { commands } from './commands';
