+++ packages/playwright-core/src/tools/trace/traceCli.ts
@@ -14,16 +14,6 @@
  * limitations under the License.
  */
 
-import { traceOpen } from './traceOpen';
-import { traceActions, traceAction } from './traceActions';
-import { traceRequests, traceRequest } from './traceRequests';
-import { traceConsole } from './traceConsole';
-import { traceErrors } from './traceErrors';
-import { traceSnapshot } from './traceSnapshot';
-import { traceScreenshot } from './traceScreenshot';
-import { traceAttachments, traceAttachment } from './traceAttachments';
-import { installSkill } from './installSkill';
-
 import type { Command } from '../../utilsBundle';
 
 export function addTraceCommands(program: Command, logErrorAndExit: (e: Error) => void) {
@@ -35,6 +25,7 @@ export function addTraceCommands(program: Command, logErrorAndExit: (e: Error) =
       .command('open <trace>')
       .description('extract trace file for inspection')
       .action((trace: string) => {
+        const { traceOpen } = require('./traceOpen');
         traceOpen(trace).catch(logErrorAndExit);
       });
 
@@ -44,6 +35,7 @@ export function addTraceCommands(program: Command, logErrorAndExit: (e: Error) =
       .option('--grep <pattern>', 'filter actions by title pattern')
       .option('--errors-only', 'only show failed actions')
       .action((options: { grep?: string, errorsOnly?: boolean }) => {
+        const { traceActions } = require('./traceActions');
         traceActions(options).catch(logErrorAndExit);
       });
 
@@ -51,6 +43,7 @@ export function addTraceCommands(program: Command, logErrorAndExit: (e: Error) =
       .command('action <action-id>')
       .description('show details of a specific action')
       .action((actionId: string) => {
+        const { traceAction } = require('./traceActions');
         traceAction(actionId).catch(logErrorAndExit);
       });
 
@@ -62,6 +55,7 @@ export function addTraceCommands(program: Command, logErrorAndExit: (e: Error) =
       .option('--status <code>', 'filter by status code')
       .option('--failed', 'only show failed requests (status >= 400)')
       .action((options: { grep?: string, method?: string, status?: string, failed?: boolean }) => {
+        const { traceRequests } = require('./traceRequests');
         traceRequests(options).catch(logErrorAndExit);
       });
 
@@ -69,6 +63,7 @@ export function addTraceCommands(program: Command, logErrorAndExit: (e: Error) =
       .command('request <request-id>')
       .description('show details of a specific network request')
       .action((requestId: string) => {
+        const { traceRequest } = require('./traceRequests');
         traceRequest(requestId).catch(logErrorAndExit);
       });
 
@@ -80,6 +75,7 @@ export function addTraceCommands(program: Command, logErrorAndExit: (e: Error) =
       .option('--browser', 'only browser console messages')
       .option('--stdio', 'only stdout/stderr')
       .action((options: { errorsOnly?: boolean, warnings?: boolean, browser?: boolean, stdio?: boolean }) => {
+        const { traceConsole } = require('./traceConsole');
         traceConsole(options).catch(logErrorAndExit);
       });
 
@@ -87,6 +83,7 @@ export function addTraceCommands(program: Command, logErrorAndExit: (e: Error) =
       .command('errors')
       .description('show errors with stack traces')
       .action(() => {
+        const { traceErrors } = require('./traceErrors');
         traceErrors().catch(logErrorAndExit);
       });
 
@@ -99,6 +96,7 @@ export function addTraceCommands(program: Command, logErrorAndExit: (e: Error) =
       .allowExcessArguments(true)
       .action(async (actionId: string, options: { name?: string, serve?: boolean }, cmd: Command) => {
         try {
+          const { traceSnapshot } = require('./traceSnapshot');
           // Collect everything after '--' as the browser command.
           const browserArgs = cmd.args.slice(1);
           await traceSnapshot(actionId, { ...options, browserArgs });
@@ -112,6 +110,7 @@ export function addTraceCommands(program: Command, logErrorAndExit: (e: Error) =
       .description('save screencast screenshot for an action')
       .option('-o, --output <path>', 'output file path')
       .action((actionId: string, options: { output?: string }) => {
+        const { traceScreenshot } = require('./traceScreenshot');
         traceScreenshot(actionId, options).catch(logErrorAndExit);
       });
 
@@ -119,6 +118,7 @@ export function addTraceCommands(program: Command, logErrorAndExit: (e: Error) =
       .command('attachments')
       .description('list trace attachments')
       .action(() => {
+        const { traceAttachments } = require('./traceAttachments');
         traceAttachments().catch(logErrorAndExit);
       });
 
@@ -127,6 +127,7 @@ export function addTraceCommands(program: Command, logErrorAndExit: (e: Error) =
       .description('extract a trace attachment by its number')
       .option('-o, --output <path>', 'output file path')
       .action((attachmentId: string, options: { output?: string }) => {
+        const { traceAttachment } = require('./traceAttachments');
         traceAttachment(attachmentId, options).catch(logErrorAndExit);
       });
 
@@ -134,6 +135,7 @@ export function addTraceCommands(program: Command, logErrorAndExit: (e: Error) =
       .command('install-skill')
       .description('install SKILL.md for LLM integration')
       .action(() => {
+        const { installSkill } = require('./installSkill');
         installSkill().catch(logErrorAndExit);
       });
 }
