+++ packages/playwright-core/src/tools/utils/mcp/DEPS.list
@@ -1,4 +1,5 @@
 [*]
 ../../../mcpBundle.ts
+../../../zodBundle.ts
 ../../../utilsBundle.ts
 ../../../server/utils/network.ts
