+++ packages/playwright/src/mcp/test/testTools.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { z } from 'playwright-core/lib/mcpBundle';
+import { z } from 'playwright-core/lib/zodBundle';
 import ListModeReporter from '../../reporters/listModeReporter';
 import { defineTestTool } from './testTool';
